+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - Allow declaration of multiple `Serializers` in the same file.
 - Improve error message for field without type.
 - Fix name of classes generated from a private class.
+- Add `serialize` field to `@BuiltValueField`. Use this to tag fields to skip
+  when serializing.
 
 ## 4.4.1